requirements.txt
pandas>=2.0.0
hydrobr
azure-storage-blob
matplotlib
python-dotenv